requirements.txt
djangorestframework>=2.4.3
